+++ R/clique_community_names.R
@@ -25,6 +25,9 @@
 #' @importFrom magrittr %>%
 
 clique_community_names <- function(g, k = 3) {
+  if (!igraph::is.igraph(g)) {
+    stop(paste0("Please use an igraph object as input. The current object is an ", class(g), "."))
+  }
   if (k > igraph::clique_num(g)) {
     stop(paste0("The maximum clique size in the network is ", igraph::clique_num(g), ". Therefore k cannot exceed this number"))
   }