+++ R/cyto_create_graph.R
@@ -33,6 +33,9 @@ cyto_create_graph <- function(graph_input,
     message("Cytoscape is not running, please start Cytoscape first")
     stop()
   }
+  if (!igraph::is.igraph(graph_input)) {
+    stop(paste0("Please use an igraph object as input. The current object is an ", class(graph_input), "."))
+  }
   varnames <- all.vars(match.call())
   for (v in varnames) {
     if (!exists(v)) {